requirements.txt
astroid==2.4.1
autopep8==1.5.2
beautifulsoup4==4.9.1
CacheControl==0.12.6
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
decorator==4.4.2
firebase-admin==4.3.0
future==0.18.2
geocoder==1.38.1
google-api-core==1.17.0
google-api-python-client==1.8.3
google-auth==1.14.3
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.28.1
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
googlemaps==4.4.0
grpcio==1.29.0
html5lib==1.0.1
httplib2==0.17.3
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.5.0
mccabe==0.6.1
msgpack==1.0.0
mysql-connector-python==8.0.20
numpy==1.18.4
pandas==1.0.3
protobuf==3.12.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pygeocoder==1.2.5
pylint==2.5.2
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
ratelim==0.1.6
requests==2.23.0
rsa==4.0
six==1.14.0
soupsieve==2.0.1
tmdbsimple==2.2.2
tmdbv3api==1.5.1
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.9
webencodings==0.5.1
wrapt==1.12.1
